+++ lib/t_db.dart
@@ -6,4 +6,3 @@ library;
 export 'src/t_db_base.dart';
 export 'src/core/type/index.dart';
 
-// TODO: Export any libraries intended for clients of this package.